Refactor repository structure and enhance configuration.
Migrated repositories to dedicated JPA and MongoDB packages for improved organization and clarity. Introduced new configurations for asynchronous execution, scheduling, and repository management. Simplified application entry point by moving bean definitions to specialized configuration classes.
