+++ agrirouter-middleware-application/src/main/java/de/agrirouter/middleware/config/AsyncConfiguration.java
@@ -1,0 +1,49 @@
+package de.agrirouter.middleware.config;
+
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.scheduling.annotation.EnableAsync;
+import org.springframework.scheduling.concurrent.ThreadPoolTaskExecutor;
+
+import java.util.concurrent.LinkedBlockingQueue;
+import java.util.concurrent.ThreadPoolExecutor;
+import java.util.concurrent.TimeUnit;
+
+@Slf4j
+@EnableAsync
+@Configuration
+public class AsyncConfiguration {
+
+    private static final int CORE_POOL_SIZE = 5;
+    private static final int MAX_POOL_SIZE = 10;
+    private static final int QUEUE_CAPACITY = 25;
+
+    /**
+     * Create a new instance of the thread pool task executor.
+     *
+     * @return -
+     */
+    @Bean
+    public ThreadPoolTaskExecutor threadPoolTaskExecutor() {
+        log.info("Creating thread pool executor, using a core pool size of 5, a max pool size of 10, and a queue capacity of 25.");
+        var taskExecutor = new ThreadPoolTaskExecutor();
+        taskExecutor.setCorePoolSize(CORE_POOL_SIZE);
+        taskExecutor.setMaxPoolSize(MAX_POOL_SIZE);
+        taskExecutor.setQueueCapacity(QUEUE_CAPACITY);
+        return taskExecutor;
+    }
+
+    /**
+     * Create a new instance of the thread pool executor.
+     *
+     * @return -
+     */
+    @Bean
+    public ThreadPoolExecutor threadPoolExecutor() {
+        var threadPoolExecutor = new ThreadPoolExecutor(CORE_POOL_SIZE, MAX_POOL_SIZE, 0L, TimeUnit.MILLISECONDS, new LinkedBlockingQueue<>(QUEUE_CAPACITY));
+        log.info("Creating thread pool executor, using a core pool size of 5, a max pool size of 10, a queue capacity of 25 and a keep alive time of o.");
+        return threadPoolExecutor;
+    }
+
+}
