+++ agrirouter-middleware-application/src/main/java/de/agrirouter/middleware/config/JpaRepositoriesConfiguration.java
@@ -1,0 +1,10 @@
+package de.agrirouter.middleware.config;
+
+import org.springframework.context.annotation.Configuration;
+
+/**
+ * Configuration for JPA repositories.
+ */
+@Configuration
+public class JpaRepositoriesConfiguration {
+}
