+++ agrirouter-middleware-application/src/main/java/de/agrirouter/middleware/config/SchedulingConfiguration.java
@@ -1,0 +1,9 @@
+package de.agrirouter.middleware.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.scheduling.annotation.EnableScheduling;
+
+@Configuration
+@EnableScheduling
+public class SchedulingConfiguration {
+}
