+++ agrirouter-middleware-business/src/main/java/de/agrirouter/middleware/business/ApplicationService.java
@@ -10,9 +10,9 @@ import de.agrirouter.middleware.api.logging.ApplicationLogInformation;
 import de.agrirouter.middleware.api.logging.BusinessOperationLogService;
 import de.agrirouter.middleware.business.parameters.AddRouterDeviceParameters;
 import de.agrirouter.middleware.domain.*;
-import de.agrirouter.middleware.persistence.ApplicationRepository;
-import de.agrirouter.middleware.persistence.RouterDeviceRepository;
-import de.agrirouter.middleware.persistence.TenantRepository;
+import de.agrirouter.middleware.persistence.jpa.ApplicationRepository;
+import de.agrirouter.middleware.persistence.jpa.RouterDeviceRepository;
+import de.agrirouter.middleware.persistence.jpa.TenantRepository;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.scheduling.annotation.Async;
