+++ agrirouter-middleware-business/src/main/java/de/agrirouter/middleware/business/DeviceDescriptionService.java
@@ -21,8 +21,8 @@ import de.agrirouter.middleware.domain.documents.Device;
 import de.agrirouter.middleware.domain.documents.DeviceDescription;
 import de.agrirouter.middleware.integration.SendMessageIntegrationService;
 import de.agrirouter.middleware.integration.parameters.MessagingIntegrationParameters;
-import de.agrirouter.middleware.persistence.DeviceDescriptionRepository;
-import de.agrirouter.middleware.persistence.DeviceRepository;
+import de.agrirouter.middleware.persistence.mongo.DeviceDescriptionRepository;
+import de.agrirouter.middleware.persistence.mongo.DeviceRepository;
 import de.saschadoemer.iso11783.clientname.ClientName;
 import de.saschadoemer.iso11783.clientname.ClientNameDecoder;
 import efdi.GrpcEfdi;
