+++ agrirouter-middleware-business/src/main/java/de/agrirouter/middleware/business/RemoveEndpointDataService.java
@@ -3,7 +3,9 @@ package de.agrirouter.middleware.business;
 import de.agrirouter.middleware.business.cache.endpoints.InternalEndpointCache;
 import de.agrirouter.middleware.domain.Endpoint;
 import de.agrirouter.middleware.integration.mqtt.MqttClientManagementService;
-import de.agrirouter.middleware.persistence.*;
+import de.agrirouter.middleware.persistence.jpa.*;
+import de.agrirouter.middleware.persistence.mongo.DeviceDescriptionRepository;
+import de.agrirouter.middleware.persistence.mongo.TimeLogRepository;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Service;
