+++ agrirouter-middleware-business/src/main/java/de/agrirouter/middleware/business/cache/messaging/MessageCache.java
@@ -3,7 +3,7 @@ package de.agrirouter.middleware.business.cache.messaging;
 import de.agrirouter.middleware.business.events.ResendMessageCacheEntryEvent;
 import de.agrirouter.middleware.domain.documents.MessageCacheEntry;
 import de.agrirouter.middleware.integration.parameters.MessagingIntegrationParameters;
-import de.agrirouter.middleware.persistence.MessageCacheEntryRepository;
+import de.agrirouter.middleware.persistence.mongo.MessageCacheEntryRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.ThreadUtils;
