+++ agrirouter-middleware-persistence/src/main/java/de/agrirouter/middleware/persistence/jpa/ApplicationRepository.java
@@ -1,4 +1,4 @@
-package de.agrirouter.middleware.persistence;
+package de.agrirouter.middleware.persistence.jpa;
 
 import de.agrirouter.middleware.domain.Application;
 import de.agrirouter.middleware.domain.Endpoint;
