+++ agrirouter-middleware-persistence/src/main/java/de/agrirouter/middleware/persistence/mongo/TimeLogRepository.java
@@ -1,4 +1,4 @@
-package de.agrirouter.middleware.persistence;
+package de.agrirouter.middleware.persistence.mongo;
 
 import de.agrirouter.middleware.domain.documents.TimeLog;
 import org.springframework.data.mongodb.repository.MongoRepository;
